moved target from SmoothControls to controls
