+++ packages/model-viewer/src/features/controls.ts
@@ -17,14 +17,17 @@ import {property} from 'lit-element';
 import {Event, PerspectiveCamera, Spherical, Vector3} from 'three';
 
 import {style} from '../decorators.js';
-import ModelViewerElementBase, {$ariaLabel, $container, $loadedTime, $needsRender, $onModelLoad, $onResize, $scene, $tick, $userInputElement, Vector3D} from '../model-viewer-base.js';
+import ModelViewerElementBase, {$ariaLabel, $container, $loadedTime, $needsRender, $onModelLoad, $onResize, $renderer, $scene, $tick, $userInputElement, Vector3D} from '../model-viewer-base.js';
 import {degreesToRadians, normalizeUnit} from '../styles/conversions.js';
 import {EvaluatedStyle, Intrinsics, SphericalIntrinsics, StyleEvaluator, Vector3Intrinsics} from '../styles/evaluators.js';
 import {IdentNode, NumberNode, numberNode, parseExpressions} from '../styles/parsers.js';
+import {Damper} from '../three-components/Damper.js';
 import {SAFE_RADIUS_RATIO} from '../three-components/Model.js';
 import {ChangeEvent, ChangeSource, PointerChangeEvent, SmoothControls} from '../three-components/SmoothControls.js';
 import {Constructor} from '../utilities.js';
 import {timeline} from '../utilities/animation.js';
+
+
 
 // NOTE(cdata): The following "animation" timing functions are deliberately
 // being used in favor of CSS animations. In Safari 12.1 and 13, CSS animations
@@ -213,6 +216,11 @@ const $focusedTime = Symbol('focusedTime');
 const $zoomAdjustedFieldOfView = Symbol('zoomAdjustedFieldOfView');
 const $lastSpherical = Symbol('lastSpherical');
 const $jumpCamera = Symbol('jumpCamera');
+const $target = Symbol('target');
+const $goalTarget = Symbol('goalTarget');
+const $targetDamperX = Symbol('targetDamperX');
+const $targetDamperY = Symbol('targetDamperY');
+const $targetDamperZ = Symbol('targetDamperZ');
 
 const $syncCameraOrbit = Symbol('syncCameraOrbit');
 const $syncFieldOfView = Symbol('syncFieldOfView');
@@ -340,6 +348,11 @@ export const ControlsMixin = <T extends Constructor<ModelViewerElementBase>>(
     protected[$zoomAdjustedFieldOfView] = 0;
     protected[$lastSpherical] = new Spherical();
     protected[$jumpCamera] = false;
+    private[$target] = new Vector3();
+    private[$goalTarget] = new Vector3();
+    private[$targetDamperX] = new Damper();
+    private[$targetDamperY] = new Damper();
+    private[$targetDamperZ] = new Damper();
 
     protected[$changeHandler] = (event: Event) =>
         this[$onChange](event as ChangeEvent);
@@ -356,7 +369,7 @@ export const ControlsMixin = <T extends Constructor<ModelViewerElementBase>>(
     }
 
     getCameraTarget(): Vector3D {
-      return this[$controls].getTarget();
+      return this[$target];
     }
 
     getFieldOfView(): number {
@@ -447,6 +460,9 @@ export const ControlsMixin = <T extends Constructor<ModelViewerElementBase>>(
       if (this[$jumpCamera] === true) {
         Promise.resolve().then(() => {
           this[$controls].jumpToGoal();
+          const goal = this[$goalTarget];
+          this[$target].copy(goal);
+          this[$scene].setTarget(goal.x, goal.y, goal.z);
           this[$jumpCamera] = false;
         });
       }
@@ -489,12 +505,15 @@ export const ControlsMixin = <T extends Constructor<ModelViewerElementBase>>(
 
     [$syncCameraTarget](style: EvaluatedStyle<Vector3Intrinsics>) {
       const [x, y, z] = style;
-      const scene = this[$scene];
-      scene.setTarget(x, y, z);
+      this[$goalTarget].set(x, y, z);
     }
 
     [$tick](time: number, delta: number) {
       super[$tick](time, delta);
+
+      if (this[$renderer].isPresenting) {
+        return;
+      }
 
       const now = performance.now();
       if (this[$waitingToPromptUser] &&
@@ -543,6 +562,17 @@ export const ControlsMixin = <T extends Constructor<ModelViewerElementBase>>(
 
         this[$lastPromptOffset] = offset;
         this[$needsRender]();
+      }
+
+      const goal = this[$goalTarget];
+      if (!goal.equals(this[$target])) {
+        const radius = this[$scene].model.idealCameraDistance;
+        let {x, y, z} = this[$target];
+        x = this[$targetDamperX].update(x, goal.x, delta, radius);
+        y = this[$targetDamperY].update(y, goal.y, delta, radius);
+        z = this[$targetDamperZ].update(z, goal.z, delta, radius);
+        this[$target].set(x, y, z);
+        this[$scene].setTarget(x, y, z);
       }
 
       this[$controls].update(time, delta);
@@ -613,6 +643,10 @@ export const ControlsMixin = <T extends Constructor<ModelViewerElementBase>>(
     }
 
     [$onResize](event: any) {
+      if (this[$renderer].isPresenting) {
+        return;
+      }
+
       const controls = this[$controls];
       const oldFramedFieldOfView = this[$scene].framedFieldOfView;
