+++ packages/model-viewer/src/three-components/Renderer.ts
@@ -257,7 +257,7 @@ export class Renderer extends EventDispatcher {
     }
 
     for (const scene of this.scenes) {
-      if (!scene.visible || scene.paused) {
+      if (!scene.element.modelIsVisible || scene.paused) {
         continue;
       }
